refactor: lib folder
