+++ angular.json
@@ -117,38 +117,38 @@
         }
       }
     },
-    "@saulmoro/transloco-localize-router": {
+    "transloco-localize-router": {
       "projectType": "library",
-      "root": "projects/saulmoro/transloco-localize-router",
-      "sourceRoot": "projects/saulmoro/transloco-localize-router/src",
+      "root": "projects/transloco-localize-router",
+      "sourceRoot": "projects/transloco-localize-router/src",
       "prefix": "lib",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:ng-packagr",
           "options": {
-            "tsConfig": "projects/saulmoro/transloco-localize-router/tsconfig.lib.json",
-            "project": "projects/saulmoro/transloco-localize-router/ng-package.json"
+            "tsConfig": "projects/transloco-localize-router/tsconfig.lib.json",
+            "project": "projects/transloco-localize-router/ng-package.json"
           },
           "configurations": {
             "production": {
-              "tsConfig": "projects/saulmoro/transloco-localize-router/tsconfig.lib.prod.json"
+              "tsConfig": "projects/transloco-localize-router/tsconfig.lib.prod.json"
             }
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/saulmoro/transloco-localize-router/src/test.ts",
-            "tsConfig": "projects/saulmoro/transloco-localize-router/tsconfig.spec.json",
-            "karmaConfig": "projects/saulmoro/transloco-localize-router/karma.conf.js"
+            "main": "projects/transloco-localize-router/src/test.ts",
+            "tsConfig": "projects/transloco-localize-router/tsconfig.spec.json",
+            "karmaConfig": "projects/transloco-localize-router/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "projects/saulmoro/transloco-localize-router/tsconfig.lib.json",
-              "projects/saulmoro/transloco-localize-router/tsconfig.spec.json"
+              "projects/transloco-localize-router/tsconfig.lib.json",
+              "projects/transloco-localize-router/tsconfig.spec.json"
             ],
             "exclude": ["**/node_modules/**"]
           }
