--- projects/saulmoro/transloco-localize-router/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../dist/saulmoro/transloco-localize-router",
-  "lib": {
-    "entryFile": "src/public-api.ts"
-  }
-}
