+++ projects/transloco-localize-router/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/transloco-localize-router",
+  "lib": {
+    "entryFile": "src/public-api.ts"
+  }
+}
