+++ projects/transloco-localize-router/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@saulmoro/transloco-localize-router",
+  "name": "transloco-localize-router",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^11.0.0",
