+++ projects/transloco-localize-router/src/lib/transloco-localize-router.config.ts
