+++ projects/transloco-localize-router/src/lib/transloco-localize-router.pipe.ts
