+++ projects/transloco-localize-router/tsconfig.lib.json
@@ -1,8 +1,8 @@
 /* To learn more about this file see: https://angular.io/config/tsconfig. */
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../../out-tsc/lib",
+    "outDir": "../../out-tsc/lib",
     "target": "es2015",
     "declaration": true,
     "declarationMap": true,
