+++ projects/transloco-localize-router/tsconfig.lib.prod.json
