+++ projects/transloco-localize-router/tsconfig.spec.json
@@ -1,8 +1,8 @@
 /* To learn more about this file see: https://angular.io/config/tsconfig. */
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../../out-tsc/spec",
+    "outDir": "../../out-tsc/spec",
     "types": ["jasmine"]
   },
   "files": ["src/test.ts"],
