+++ projects/transloco-localize-router/tslint.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tslint.json",
+  "extends": "../../tslint.json",
   "rules": {
     "directive-selector": [true, "attribute", "lib", "camelCase"],
     "component-selector": [true, "element", "lib", "kebab-case"]
