+++ src/app/core/layout/layout.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { TranslocoLocalizeRouterModule } from '@saulmoro/transloco-localize-router';
+import { TranslocoLocalizeRouterModule } from 'transloco-localize-router';
 
 import { HeaderComponent } from './header/header.component';
 import { LanguageDropdownComponent } from './header/language-dropdown/language-dropdown.component';
